+++ wagtail/snippets/views/snippets.py
@@ -13,6 +13,7 @@ from wagtail.admin import messages
 from wagtail.admin.edit_handlers import ObjectList, extract_panel_definitions_from_model_class
 from wagtail.admin.forms.search import SearchForm
 from wagtail.admin.utils import permission_denied
+from wagtail.core.collectors import get_paginated_uses
 from wagtail.search.backends import get_search_backend
 from wagtail.search.index import class_is_indexed
 from wagtail.snippets.models import get_snippet_models
@@ -234,7 +235,9 @@ def delete(request, app_label, model_name, pk=None):
 
     count = len(instances)
 
-    if request.method == 'POST':
+    uses = get_paginated_uses(request, *instances)
+
+    if request.method == 'POST' and not uses.are_protected:
         for instance in instances:
             instance.delete()
 
@@ -261,17 +264,5 @@ def delete(request, app_label, model_name, pk=None):
             reverse('wagtailsnippets:delete-multiple', args=(app_label, model_name))
             + '?' + urlencode([('id', instance.pk) for instance in instances])
         ),
-    })
-
-
-def usage(request, app_label, model_name, pk):
-    model = get_snippet_model_from_url_params(app_label, model_name)
-    instance = get_object_or_404(model, pk=unquote(pk))
-
-    paginator = Paginator(instance.get_usage(), per_page=20)
-    used_by = paginator.get_page(request.GET.get('p'))
-
-    return render(request, "wagtailsnippets/snippets/usage.html", {
-        'instance': instance,
-        'used_by': used_by
+        'uses': uses,
     })
